test(2d gldm): rename the PyRadiomics pytest to the mechanics guard it is
tests/python/test_2d_gldm_pyradiomics.py asserted three GLDM features against numbers its
docstring called "the PyRadiomics reference", at rel=0.20 and rel=0.15.

The numbers reproduce -- they are genuinely PyRadiomics' output on that fixture. What does not
hold is the comparison. The test ran Nyxus at the production default (ibsi=false, MATLAB grey
binning at coarse_gray_depth=64) against a PyRadiomics run binning at binWidth=1, so the two
built their dependence matrices over different level assignments. Measured across all 14
features on that fixture the disagreement runs from 1.3% to 108% (GLDM_SDLGLE), and the bands
were wide enough to pass the three assertions anyway: GLDM_LDE sat 8.8% from its "reference".
SPEC 5 does not let a tolerance absorb a configuration mismatch, so this was not an oracle
test at any band.

Its real subject is the bug #14b background-pollution guard on the production path, which the
IBSI-mode oracle tests cannot reach because they run on fully-masked ROIs. That is a SPEC 2
mechanics test, and its C++ counterpart test_2d_gldm_mechanics.h already carried the right
name and pointed at this file under a third one. Renamed, and re-pinned against Nyxus' own
values at rel=1e-9 for the five count and dependence features the bug moved, keeping the
GLDM_LDE < 5.0 bound that fails on the pre-fix inflated ~16.6.

not_covered.md records the variant: an oracle-named test whose values are genuinely the tool's,
asserted at a band far wider than the tool's own precision. The width is the tell.
